+++ tests/testthat/test-get_sysreqs.R
@@ -1,5 +1,0 @@
-test_that("get_sysreqs works", {
-	expect_warning(
-		get_sysreqs()
-	)
-})